web_scrapers/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
python-decouple==3.4
CurrencyConverter==0.16.1